+++ backend/canisters/group_index/api/src/lifecycle/init.rs
@@ -13,6 +13,7 @@ pub struct Args {
     pub registry_canister_id: CanisterId,
     pub internet_identity_canister_id: CanisterId,
     pub video_call_operators: Vec<Principal>,
+    pub rng_seed: [u8; 32],
     pub wasm_version: BuildVersion,
     pub test_mode: bool,
 }