+++ backend/canisters/identity/api/src/lifecycle/init.rs
@@ -10,6 +10,7 @@ pub struct Args {
     pub sign_in_with_email_canister_id: CanisterId,
     pub originating_canisters: Vec<CanisterId>,
     pub skip_captcha_whitelist: Vec<CanisterId>,
+    pub rng_seed: [u8; 32],
     pub wasm_version: BuildVersion,
     pub test_mode: bool,
 }