+++ backend/canisters/local_user_index/impl/src/lifecycle/init.rs
@@ -16,7 +16,7 @@ fn init(args: Args) {
     stable_memory_map::init(get_stable_memory_map_memory());
     init_cycles_dispenser_client(args.cycles_dispenser_canister_id, args.test_mode);
 
-    let env = init_env([0; 32]);
+    let env = init_env(args.rng_seed);
     let canister_pool_target_size = if args.test_mode { 3_u16 } else { CANISTER_POOL_TARGET_SIZE };
 
     let data = Data::new(