+++ backend/canisters/user/impl/src/lifecycle/init.rs
@@ -1,4 +1,4 @@
-use crate::lifecycle::{init_env, init_state};
+use crate::lifecycle::init_state;
 use crate::memory::get_stable_memory_map_memory;
 use crate::{Data, mutate_state, openchat_bot};
 use canister_tracing_macros::trace;
@@ -6,6 +6,7 @@ use ic_cdk::init;
 use tracing::info;
 use user_canister::init::Args;
 use utils::env::Environment;
+use utils::env::canister::CanisterEnv;
 
 #[init]
 #[trace]
@@ -13,7 +14,7 @@ fn init(args: Args) {
     canister_logger::init(args.test_mode);
     stable_memory_map::init(get_stable_memory_map_memory());
 
-    let env = init_env([0; 32]);
+    let env = Box::new(CanisterEnv::new(args.rng_seed));
     let now = env.now();
 
     let data = Data::new(