+++ backend/canisters/user_index/impl/src/lifecycle/init.rs
@@ -1,5 +1,5 @@
 use crate::Data;
-use crate::lifecycle::{init_env, init_state};
+use crate::lifecycle::init_state;
 use crate::memory::get_stable_memory_map_memory;
 use canister_tracing_macros::trace;
 use ic_cdk::init;
@@ -7,6 +7,7 @@ use tracing::info;
 use user_index_canister::init::Args;
 use utils::cycles::init_cycles_dispenser_client;
 use utils::env::Environment;
+use utils::env::canister::CanisterEnv;
 
 #[init]
 #[trace]
@@ -15,7 +16,7 @@ fn init(args: Args) {
     stable_memory_map::init(get_stable_memory_map_memory());
     init_cycles_dispenser_client(args.cycles_dispenser_canister_id, args.test_mode);
 
-    let env = init_env([0; 32], false);
+    let env = Box::new(CanisterEnv::new(args.rng_seed));
 
     let data = Data::new(
         args.governance_principals,
@@ -35,6 +36,7 @@ fn init(args: Args) {
         args.translations_canister_id,
         args.website_canister_id,
         args.video_call_operators,
+        args.oc_secret_key_der,
         args.test_mode,
         env.now(),
     );