+++ backend/canisters/identity/c2c_client/src/lib.rs
@@ -4,5 +4,5 @@ use identity_canister::*;
 // Queries
 
 // Updates
-generate_c2c_call!(c2c_set_user_ids);
+generate_c2c_call!(c2c_set_oc_secret_key);
 generate_c2c_call!(c2c_set_user_identities);