+++ backend/canisters/local_user_index/impl/src/lib.rs
@@ -106,7 +106,7 @@ impl RuntimeState {
                     self.data.internet_identity_canister_id,
                     self.data.website_canister_id,
                     jwt,
-                    &self.data.ic_root_key,
+                    &self.env.ic_root_key(),
                     now,
                 ) {
                     self.push_event_to_user_index(
@@ -570,8 +570,6 @@ struct Data {
     pub event_store_client: EventStoreClient<CdkRuntime>,
     pub event_deduper: EventDeduper,
     pub users_to_delete_queue: VecDeque<UserToDelete>,
-    #[serde(with = "serde_bytes")]
-    pub ic_root_key: Vec<u8>,
     pub events_for_remote_users: Vec<(UserId, UserEvent)>,
     pub cycles_balance_check_queue: VecDeque<CanisterId>,
     pub fire_and_forget_handler: FireAndForgetHandler,
@@ -615,8 +613,7 @@ impl Data {
         website_canister_id: CanisterId,
         canister_pool_target_size: u16,
         video_call_operators: Vec<Principal>,
-        oc_secret_key_der: Option<Vec<u8>>,
-        ic_root_key: Vec<u8>,
+        oc_secret_key_der: Vec<u8>,
         test_mode: bool,
     ) -> Self {
         Data {
@@ -656,15 +653,12 @@ impl Data {
             rng_seed: [0; 32],
             notification_pushers: HashSet::new(),
             video_call_operators,
-            oc_key_pair: oc_secret_key_der
-                .map(|sk| P256KeyPair::from_secret_key_der(sk).unwrap())
-                .unwrap_or_default(),
+            oc_key_pair: P256KeyPair::from_secret_key_der(oc_secret_key_der).unwrap(),
             event_store_client: EventStoreClientBuilder::new(event_relay_canister_id, CdkRuntime::default())
                 .with_flush_delay(Duration::from_millis(MINUTE_IN_MS))
                 .build(),
             event_deduper: EventDeduper::default(),
             users_to_delete_queue: VecDeque::new(),
-            ic_root_key,
             events_for_remote_users: Vec::new(),
             cycles_balance_check_queue: VecDeque::new(),
             bots: BotsMap::default(),