+++ backend/canisters/user/impl/src/lifecycle/post_upgrade.rs
@@ -1,5 +1,5 @@
 use crate::Data;
-use crate::lifecycle::{init_env, init_state};
+use crate::lifecycle::init_state;
 use crate::memory::{get_stable_memory_map_memory, get_upgrades_memory};
 use canister_api_macros::post_upgrade;
 use canister_logger::LogEntry;
@@ -7,6 +7,7 @@ use canister_tracing_macros::trace;
 use stable_memory::get_reader;
 use tracing::info;
 use user_canister::post_upgrade::Args;
+use utils::env::canister::CanisterEnv;
 
 #[post_upgrade(msgpack = true)]
 #[trace]
@@ -21,7 +22,7 @@ fn post_upgrade(args: Args) {
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
-    let env = init_env(data.rng_seed);
+    let env = Box::new(CanisterEnv::new(data.rng_seed));
     init_state(env, data, args.wasm_version);
 
     let total_instructions = ic_cdk::api::call_context_instruction_counter();