+++ backend/canisters/user_index/impl/src/lifecycle/post_upgrade.rs
@@ -1,14 +1,15 @@
-use crate::Data;
-use crate::lifecycle::{init_env, init_state};
+use crate::lifecycle::init_state;
 use crate::memory::{get_stable_memory_map_memory, get_upgrades_memory};
+use crate::{Data, read_state};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk::post_upgrade;
-use identity_canister::UserIdentity;
 use stable_memory::get_reader;
+use std::time::Duration;
 use tracing::info;
 use user_index_canister::post_upgrade::Args;
 use utils::cycles::init_cycles_dispenser_client;
+use utils::env::canister::CanisterEnv;
 
 #[post_upgrade]
 #[trace]
@@ -18,27 +19,33 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+    let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
-
-    // TODO: Remove this after the canister is upgraded
-    data.identity_canister_user_sync_queue_2
-        .extend(
-            data.identity_canister_user_sync_queue
-                .drain(..)
-                .map(|(principal, user_id)| UserIdentity {
-                    principal,
-                    user_id,
-                    email: None,
-                }),
-        );
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
-    let env = init_env(data.rng_seed, data.oc_key_pair.is_initialised());
+    let env = Box::new(CanisterEnv::new(data.rng_seed));
     init_cycles_dispenser_client(data.cycles_dispenser_canister_id, data.test_mode);
     init_state(env, data, args.wasm_version);
 
     let total_instructions = ic_cdk::api::call_context_instruction_counter();
     info!(version = %args.wasm_version, total_instructions, "Post-upgrade complete");
+
+    ic_cdk_timers::set_timer(Duration::ZERO, || {
+        ic_cdk::futures::spawn(async {
+            let (identity_canister_id, oc_secret_key_der) = read_state(|state| {
+                (
+                    state.data.identity_canister_id,
+                    state.data.oc_key_pair.secret_key_der().to_vec(),
+                )
+            });
+
+            identity_canister_c2c_client::c2c_set_oc_secret_key(
+                identity_canister_id,
+                &identity_canister::c2c_set_oc_secret_key::Args { oc_secret_key_der },
+            )
+            .await
+            .unwrap();
+        });
+    });
 }